third_party/WebKit/LayoutTests/svg/dom/SVGLength-expected.txt

This test checks the SVGLength API

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Check initial length values
PASS length.unitType is SVGLength.SVG_LENGTHTYPE_NUMBER
PASS length.value is 0
PASS length.valueInSpecifiedUnits is 0
PASS length.valueAsString is "0"

Set value to be 2px
PASS length.unitType is SVGLength.SVG_LENGTHTYPE_PX
PASS length.value is 2
PASS length.valueInSpecifiedUnits is 2
PASS length.valueAsString is "2px"

Check invalid arguments for 'convertToSpecifiedUnits'
PASS length.convertToSpecifiedUnits(SVGLength.SVG_LENGTHTYPE_UNKNOWN) threw exception NotSupportedError: Failed to execute 'convertToSpecifiedUnits' on 'SVGLength': Cannot convert to unknown or invalid units (0)..
PASS length.convertToSpecifiedUnits(-1) threw exception NotSupportedError: Failed to execute 'convertToSpecifiedUnits' on 'SVGLength': Cannot convert to unknown or invalid units (65535)..
PASS length.convertToSpecifiedUnits(11) threw exception NotSupportedError: Failed to execute 'convertToSpecifiedUnits' on 'SVGLength': Cannot convert to unknown or invalid units (11)..
PASS length.convertToSpecifiedUnits('aString') threw exception NotSupportedError: Failed to execute 'convertToSpecifiedUnits' on 'SVGLength': Cannot convert to unknown or invalid units (0)..
PASS length.convertToSpecifiedUnits(length) threw exception NotSupportedError: Failed to execute 'convertToSpecifiedUnits' on 'SVGLength': Cannot convert to unknown or invalid units (0)..
PASS length.convertToSpecifiedUnits(svgElement) threw exception NotSupportedError: Failed to execute 'convertToSpecifiedUnits' on 'SVGLength': Cannot convert to unknown or invalid units (0)..
PASS length.convertToSpecifiedUnits() threw exception TypeError: Failed to execute 'convertToSpecifiedUnits' on 'SVGLength': 1 argument required, but only 0 present..
PASS length.unitType is SVGLength.SVG_LENGTHTYPE_PX
PASS length.value is 2
PASS length.valueInSpecifiedUnits is 2
PASS length.unitType is SVGLength.SVG_LENGTHTYPE_PX

Check invalid arguments for 'newValueSpecifiedUnits'
PASS length.newValueSpecifiedUnits(SVGLength.SVG_LENGTHTYPE_UNKNOWN, 4) threw exception NotSupportedError: Failed to execute 'newValueSpecifiedUnits' on 'SVGLength': Cannot set value with unknown or invalid units (0)..
PASS length.newValueSpecifiedUnits(-1, 4) threw exception NotSupportedError: Failed to execute 'newValueSpecifiedUnits' on 'SVGLength': Cannot set value with unknown or invalid units (65535)..
PASS length.newValueSpecifiedUnits(11, 4) threw exception NotSupportedError: Failed to execute 'newValueSpecifiedUnits' on 'SVGLength': Cannot set value with unknown or invalid units (11)..
PASS length.newValueSpecifiedUnits(SVGLength.SVG_LENGTHTYPE_PX, 'aString') is undefined.
PASS length.value is NaN
PASS length.newValueSpecifiedUnits(SVGLength.SVG_LENGTHTYPE_PX, 0) is undefined.
PASS length.newValueSpecifiedUnits(SVGLength.SVG_LENGTHTYPE_PX, length) is undefined.
PASS length.value is NaN
PASS length.newValueSpecifiedUnits(SVGLength.SVG_LENGTHTYPE_PX, 0) is undefined.
PASS length.newValueSpecifiedUnits(SVGLength.SVG_LENGTHTYPE_PX, svgElement) is undefined.
PASS length.value is NaN
PASS length.newValueSpecifiedUnits(SVGLength.SVG_LENGTHTYPE_PX) threw exception TypeError: Failed to execute 'newValueSpecifiedUnits' on 'SVGLength': 2 arguments required, but only 1 present..
PASS length.newValueSpecifiedUnits('aString', 4) threw exception NotSupportedError: Failed to execute 'newValueSpecifiedUnits' on 'SVGLength': Cannot set value with unknown or invalid units (0)..
PASS length.newValueSpecifiedUnits(length, 4) threw exception NotSupportedError: Failed to execute 'newValueSpecifiedUnits' on 'SVGLength': Cannot set value with unknown or invalid units (0)..
PASS length.newValueSpecifiedUnits(svgElement, 4) threw exception NotSupportedError: Failed to execute 'newValueSpecifiedUnits' on 'SVGLength': Cannot set value with unknown or invalid units (0)..
PASS length.newValueSpecifiedUnits('aString', 'aString') threw exception NotSupportedError: Failed to execute 'newValueSpecifiedUnits' on 'SVGLength': Cannot set value with unknown or invalid units (0)..
PASS length.newValueSpecifiedUnits(length, length) threw exception NotSupportedError: Failed to execute 'newValueSpecifiedUnits' on 'SVGLength': Cannot set value with unknown or invalid units (0)..
PASS length.newValueSpecifiedUnits(svgElement, svgElement) threw exception NotSupportedError: Failed to execute 'newValueSpecifiedUnits' on 'SVGLength': Cannot set value with unknown or invalid units (0)..
PASS length.unitType is SVGLength.SVG_LENGTHTYPE_PX
PASS length.value is 2
PASS length.valueInSpecifiedUnits is 2
PASS length.unitType is SVGLength.SVG_LENGTHTYPE_PX

Check setting invalid 'valueAsString' arguments
PASS length.valueAsString = '10deg' threw exception SyntaxError: Failed to set the 'valueAsString' property on 'SVGLength': The value provided ('10deg') is invalid..
PASS length.valueAsString is "2px"
PASS length.value is 2
PASS length.valueInSpecifiedUnits is 2
PASS length.unitType is SVGLength.SVG_LENGTHTYPE_PX
PASS length.valueAsString = '1pX' threw exception SyntaxError: Failed to set the 'valueAsString' property on 'SVGLength': The value provided ('1pX') is invalid..
PASS length.valueAsString is "2px"
PASS length.value is 2
PASS length.valueInSpecifiedUnits is 2
PASS length.unitType is SVGLength.SVG_LENGTHTYPE_PX
PASS length.valueAsString = ',5 em' threw exception SyntaxError: Failed to set the 'valueAsString' property on 'SVGLength': The value provided (',5 em') is invalid..
PASS length.valueAsString is "2px"
PASS length.value is 2
PASS length.valueInSpecifiedUnits is 2
PASS length.unitType is SVGLength.SVG_LENGTHTYPE_PX
PASS successfullyParsed is true

TEST COMPLETE
